requirements.txt
langchain-openai
python-dotenv
langchain
pandas
langchain-community
beautifulsoup4
openpyxl
